add remove track

+++ app/src/main/java/com/example/playlistmaker/mvvm/audioplayer/data/dto/TrackDto.kt
@@ -1,7 +1,7 @@
 package com.example.playlistmaker.mvvm.audioplayer.data.dto
 
 data class TrackDto(
-    val trackId: Int,
+    val trackId: Long,
     val trackName: String,
     val artistName: String,
     val trackTimeMillis: Int,
